Do not install unsupported package on ARM64

+++ install/desktop/optional/app-zoom.sh
@@ -1,6 +1,8 @@
 # Make video calls using https://zoom.us/
-cd /tmp
-wget https://zoom.us/client/latest/zoom_amd64.deb
-sudo apt install -y ./zoom_amd64.deb
-rm zoom_amd64.deb
-cd -
+if [ $(dpkg --print-architecture) != "arm64" ]; then
+    cd /tmp
+    wget https://zoom.us/client/latest/zoom_amd64.deb
+    sudo apt install -y ./zoom_amd64.deb
+    rm zoom_amd64.deb
+    cd -
+fi
